+++ CHANGELOG.md
@@ -9,6 +9,7 @@ For information on changes in released versions of Teku, see the [releases page]
 ## Unreleased Changes
 
 ### Breaking Changes
+- Removed `epoch` from the POST `/eth/v1/validator/liveness/{epoch}` response as it was not part of the api spec.
 
 ### Additions and Improvements
 - The latest version of [blst](https://github.com/supranational/blst) will automatically use optimized code paths if they are supported. As a result, `JAVA_OPTS="-Dteku.portableBlst=true"` is no longer necessary for some older systems.